chore: further refactoring of funcs from handle-delimiter-arith into misc-handle-delimeter-arith and assign-implies-entries
